requirements.txt
streamlit-extras==0.4.0
streamlit-authenticator==0.5.0
PyYAML==5.4.1
fuzzywuzzy==0.18.0
pandas==1.3.3
pdf2docx==0.8.1
openai==0.11.3
gradio==2.3.7
retrying==1.3.3
flask==2.0.2
werkzeug==2.0.2
xwltwriter==1.3.8
opencv-python==4.5.3.56
Pillow==8.3.2
pytesseract==0.3.8
opencv-python-headless==4.5.3.56
textblob==0.15.3
